requirements.txt
albumentations==1.0.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.1
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
bcrypt==4.0.1
# beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
bleach==5.0.1
boto3==1.26.30
botocore==1.29.30
brotlipy==0.7.0
certifi==2022.12.7
# cffi @ file:///tmp/build/80754af9/cffi_1605538068321/work
# chardet @ file:///tmp/build/80754af9/chardet_1605303185383/work
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
conda==4.3.16
conda-build==2.1.5
# conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1603018141399/work
contourpy==1.0.6
# cryptography @ file:///tmp/build/80754af9/cryptography_1605544487601/work
cycler==0.11.0
Cython==0.29.32
debugpy==1.6.4
decorator==4.4.2
defusedxml==0.7.1
dnspython==2.1.0
docker-pycreds==0.4.0
docopt==0.6.2
easydict==1.10
efficientnet-pytorch==0.7.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
# filelock @ file:///home/linux1/recipes/ci/filelock_1610993975404/work
fonttools==4.38.0
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.29
# glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
greenlet==2.0.1
huggingface-hub==0.11.1
# idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.22.4
importlib-metadata==5.1.0
importlib-resources==5.10.1
inotify-simple==1.2.1
ipykernel==6.19.2
ipython==8.7.0
# ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==8.0.3
# jedi @ file:///tmp/build/80754af9/jedi_1606932564285/work
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.10
jsonschema==4.17.3
jupyter-client==7.4.8
jupyter-core==5.1.0
jupyter-events==0.5.0
jupyter-server==2.0.1
jupyter-server-terminals==0.4.2
jupyterlab==3.5.1
jupyterlab-pygments==0.2.2
jupyterlab-server==2.16.5
jupyterlab-widgets==3.0.4
kiwisolver==1.4.4
# libarchive-c @ file:///home/linux1/recipes/ci/python-libarchive-c_1610974153025/work
MarkupSafe==2.0.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
mmcls==0.25.0
munch==2.5.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
notebook==6.5.2
notebook-shim==0.2.2
numpy==1.19.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
olefile==0.46
opencv-python==4.5.5.64
opencv-python-headless==4.6.0.66
packaging==22.0
pandas==1.3.4
pandocfilters==1.5.0
paramiko==2.12.0
parso==0.7.0
pathtools==0.1.2
# pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
# pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.1.0
pipreqs==0.4.11
pkginfo==1.7.0
pkgutil-resolve-name==1.3.10
platformdirs==2.6.0
# pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1667232663820/work
pretrainedmodels==0.7.4
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.36
protobuf==3.19.6
# psutil @ file:///tmp/build/80754af9/psutil_1598370257551/work
# ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval==0.2.2
pycocotools==2.0.4
pycosat==0.6.3
# pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
# pydensecrf @ file:///home/conda/feedstock_root/build_artifacts/pydensecrf_1667219249912/work
# Pygments @ file:///tmp/build/80754af9/pygments_1610565767015/work
PyNaCl==1.5.0
# pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1605545627475/work
pyparsing==3.0.9
pyrsistent==0.19.2
# PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
python-etcd==0.4.5
python-json-logger==2.0.4
# pytz @ file:///tmp/build/80754af9/pytz_1608922264688/work
PyWavelets==1.4.1
PyYAML==5.3.1
pyzmq==24.0.1
requests==2.28.1
retrying==1.3.4
ruamel-yaml==0.15.87
# ruamel.yaml.clib @ file:///home/conda/feedstock_root/build_artifacts/ruamel.yaml.clib_1670412724006/work
s3transfer==0.6.0
sagemaker-training==4.4.1
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.7.3
seaborn==0.11.2
# segmentation-models-pytorch @ git+https://github.com/qubvel/segmentation_models.pytorch@478b7d2d295cbd7787f4799c687fc18e82013829
Send2Trash==1.8.0
sentry-sdk==1.12.1
setproctitle==1.3.2
shortuuid==1.0.11
# six @ file:///tmp/build/80754af9/six_1605205327372/work
smmap==5.0.0
sniffio==1.3.0
# soupsieve @ file:///tmp/build/80754af9/soupsieve_1607965878077/work
stack-data==0.6.2
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2022.10.10
timm==0.6.12
tinycss2==1.2.1
tomli==2.0.1
# toolz @ file:///home/conda/feedstock_root/build_artifacts/toolz_1657485559105/work
torch==1.13.1
torchaudio==0.7.0
torchelastic==0.2.1
torchmetrics==0.11.0
torchvision==0.8.2
tornado==6.2
# tqdm @ file:///tmp/build/80754af9/tqdm_1605303662894/work
traitlets==5.7.1
# typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1598376058250/work
urllib3==1.26.13
wandb==0.13.7
# wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webcolors==1.11.1
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.4
yarg==0.1.9
zipp==3.11.0
zope.event==4.5.0
zope.interface==5.5.2
